Plot the Clenshaw-Curtis rule
We have an offer to include the Clenshaw-Curtis rule in gluemath.  
It is well known to use the Gauss-Legendre formula to find generalized incomplete beta function.  
However, the Clenshaw-Curtis rule is easier to implement.  
Use an FFT internally. Prof. Ooura(the author) of this algorithm, has also created many patterns for FFTs.  
FFT will also be hired by Prof. Ooura.